Refactor directory structure for workers.
